Expose a tiny bit of concurrency for backbuffer rendering
By separating out flushing commands to the GPU and waiting for the
backbuffer to be rendered to and then swapping it we can possibly
recieve and handle a few events from our asynch pool before the GPU
rendering completes. Mostly we'll be able to react to window closing
and simulator updates faster.